+++ .vscode/settings.json
@@ -1,5 +1,5 @@
 {
-    "python.pythonPath": "/usr/local/bin/python3",
+    "python.pythonPath": "/usr/local/opt/python@2/bin/python2.7",
     "python.linting.pylintEnabled": true,
     "python.linting.enabled": true
 }
